Add a text stream test
